chips: qemu_arm_mps2: forbid unsafe, split it into qemu_arm_mps2-unsafe
The peripheral drivers were already handed a `StaticRef` to their own
register block, so the only `unsafe` in the shared chip crate was in code
that has to be somewhere else anyway:

  - the base addresses those `StaticRef`s are built from,
  - the `Chip` implementation, whose `sleep()` calls `wfi()` and whose
    `print_state()` is an unsafe trait method, and
  - the UART panic writer, whose `create_panic_writer()` is likewise an
    unsafe trait method.

Move those into a new `qemu_arm_mps2-unsafe`, following `nrf5x-unsafe` and
`stm32u5xx-unsafe`, and mark `qemu_arm_mps2` `#![forbid(unsafe_code)]`.

The AN385 and AN386 images have identical peripherals at identical
addresses and `QemuArmMps2Chip` is already generic over `CortexMVariant`,
so the unsafe half is shared by both images rather than duplicated into
`qemu_arm_mps2_an385`/`_an386`.

Every raw address now lives in one file, `addresses.rs`, which lists the
machine's whole memory map rather than only the peripherals currently
driven. `MPU_BASE` moves there too, though it is fixed by ARMv7-M rather
than by this machine.

`Mps2DefaultPeripherals` stays in the safe crate: given a `StaticRef` per
peripheral -- each a distinct type, so the bindings are still checked --
constructing the drivers is safe. Supplying this machine's addresses is
not, so that is `qemu_arm_mps2_unsafe::default_peripherals()`, which is
`unsafe` for the same reason `QemuArmMps2Chip::new()` is: calling it twice
would hand two drivers the same registers.

The panic writer's polling `transmit_sync()` becomes a safe method on
`Uart`, so the writer no longer needs the driver's private register
handle. This mirrors what `qemu_virt_chip` does for `Uart16550`.

No functional change.
