+++ chips/qemu_arm_mps2-unsafe/Cargo.toml
@@ -3,13 +3,15 @@
 # Copyright Tock Contributors 2026.
 
 [package]
-name = "qemu_arm_mps2"
+name = "qemu_arm_mps2-unsafe"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
 
 [dependencies]
+cortexm = { path = "../../arch/cortex-m" }
 kernel = { path = "../../kernel" }
+qemu_arm_mps2 = { path = "../qemu_arm_mps2" }
 
 [lints]
 workspace = true
