+++ chips/qemu_arm_mps2-unsafe/README.md
@@ -1,0 +1,5 @@
+ARM MPS2 "Chip" Family - Unsafe
+===============================
+
+This contains the `unsafe` code for the `qemu_arm_mps2` chip: the machine's
+memory map, the `Chip` implementation, and the panic writer.
