+++ chips/qemu_arm_mps2/src/lib.rs
@@ -2,14 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 // Copyright Tock Contributors 2026.
 
-//! Chip support for the ARM MPS2 AN385/AN386 FPGA images under QEMU.
+//! Peripheral drivers for the ARM MPS2 AN385/AN386 FPGA images under QEMU.
 //!
 //! Peripherals are common to both; each image's core and vector table are in
 //! its own crate (`qemu_arm_mps2_an385`, `qemu_arm_mps2_an386`).
 
+#![forbid(unsafe_code)]
 #![no_std]
 
-pub mod chip;
 pub mod interrupts;
 pub mod led;
 pub mod spi;
@@ -18,6 +18,7 @@ pub mod uart;
 pub mod watchdog;
 
 use kernel::platform::chip::InterruptService;
+use kernel::utilities::StaticRef;
 
 /// The MPS2 AN385/AN386 machine's fixed system clock, in Hz (`SYSCLK_FRQ`
 /// in QEMU's `hw/arm/mps2.c`), which every CMSDK peripheral's PCLK is
@@ -25,11 +26,6 @@ use kernel::platform::chip::InterruptService;
 pub const SYSCLK_FRQ: u32 = 25_000_000;
 
 /// Instantiates the peripherals this chip crate drives.
-///
-/// Only UART0, Timer0, and the "Shield0" PL022 are wired up (console/alarm
-/// backing, and the syscall-facing SPI controller); UART1-4, Timer1, and
-/// the other four PL022 instances exist on the real memory map but are
-/// unused here.
 pub struct Mps2DefaultPeripherals<'a> {
     pub uart0: uart::Uart<'a>,
     pub timer0: timer::Timer<'a>,
@@ -39,13 +35,19 @@ pub struct Mps2DefaultPeripherals<'a> {
 }
 
 impl Mps2DefaultPeripherals<'_> {
-    pub fn new() -> Self {
+    pub fn new(
+        uart0: StaticRef<uart::UartRegisters>,
+        timer0: StaticRef<timer::TimerRegisters>,
+        fpgaio: StaticRef<led::FpgaioRegisters>,
+        spi_shield0: StaticRef<spi::SpiRegisters>,
+        watchdog: StaticRef<watchdog::WatchdogRegisters>,
+    ) -> Self {
         Self {
-            uart0: uart::Uart::new(uart::UART0_BASE),
-            timer0: timer::Timer::new(timer::TIMER0_BASE),
-            fpgaio: led::Fpgaio::new(led::FPGAIO_BASE),
-            spi_shield0: spi::Spi::new(spi::SPI_SHIELD0_BASE),
-            watchdog: watchdog::Watchdog::new(watchdog::WATCHDOG_BASE),
+            uart0: uart::Uart::new(uart0),
+            timer0: timer::Timer::new(timer0),
+            fpgaio: led::Fpgaio::new(fpgaio),
+            spi_shield0: spi::Spi::new(spi_shield0),
+            watchdog: watchdog::Watchdog::new(watchdog),
         }
     }
 }
